Repackaged editor into one page to share quiz data
